pyproject.toml
[project]
name = "project3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=2.32.0",
    "opencv-python>=4.13.0.92",
    "pillow>=12.2.0",
    "python-dotenv>=1.2.2",
    "streamlit>=1.56.0",
    "tensorflow>=2.21.0",
]
